CHOICE - reminders parameter on meaningful creation of new topics
Please create a GitHub issue if you believe the README should be more clear on how to change from the "new event topic bumps 5 minutes before the evnt starts"

Please note that a location change usually causes a new topic pathway, so you get the desired bump in that case. Also in that case this commit applies.