CHOICE - reminders parameter on non-meaningful adoption of existing topics
Please create a GitHub issue if you believe the README should be more clear on how to change from the "existing event topic is modified so bumps 5 minutes before the evnt starts"

Please note that a location change usually causes a new topic pathway, so you get the desired bump in that case. This commit deos not apply to that case.